+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+* Added IP support to Requests
+* Improved environment variable handling
+* Improved x-sendfile handling
+
 ## v0.1.0 (2023-10-30)
 * Built initial implementation